refactor into one ErrorPage component with message map
