+++ packages/components/src/ErrorPage/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useMessages"
