+++ packages/components/src/ErrorPage/index.ts
@@ -1,3 +1,1 @@
 export * from "./ErrorPage"
-export { BaseErrorPage } from "./subcomponents"
-export type { BaseErrorPageProps } from "./subcomponents"
