--- packages/components/src/ErrorPage/subcomponents/index.ts
@@ -1,10 +1,0 @@
-export * from "./BaseErrorPage"
-export * from "./Error400"
-export * from "./Error401"
-export * from "./Error403"
-export * from "./Error404"
-export * from "./Error422"
-export * from "./Error500"
-export * from "./Error502"
-export * from "./Error503"
-export * from "./Error504"
