--- packages/components/src/ErrorPage/subcomponents/types.ts
@@ -1,3 +1,0 @@
-import { BaseErrorPageProps } from "./BaseErrorPage"
-
-export type ErrorProps = { callToAction?: BaseErrorPageProps["callToAction"] }
